Move shaping/analysis output type to shape module
