--- parley/src/layout/glyph.rs
@@ -1,19 +1,0 @@
-// Copyright 2025 the Parley Authors
-// SPDX-License-Identifier: Apache-2.0 OR MIT
-
-/// Glyph with an offset and advance.
-#[derive(Copy, Clone, Default, Debug, PartialEq)]
-pub struct Glyph {
-    pub id: u32,
-    pub style_index: u16,
-    pub x: f32,
-    pub y: f32,
-    pub advance: f32,
-}
-
-impl Glyph {
-    /// Returns the index into the layout style collection.
-    pub fn style_index(&self) -> usize {
-        self.style_index as usize
-    }
-}
