+++ parley/src/layout/line.rs
@@ -3,10 +3,10 @@
 
 use crate::layout::Style;
 use crate::layout::data::BreakReason;
-use crate::layout::data::{LayoutItemKind, LineData};
-use crate::layout::glyph::Glyph;
+use crate::layout::data::LineData;
 use crate::layout::layout::Layout;
 use crate::layout::run::Run;
+use crate::shape::data::{Glyph, LayoutItemKind};
 use crate::style::Brush;
 use crate::{InlineBox, InlineBoxKind};
 use core::ops::Range;
